+++ example/lib/src/controller/home/_controller.dart
@@ -1,7 +1,6 @@
 /// The export file for the directory, home.
+export 'package:example/src/controller/home/another_controller.dart';
 
 export 'package:example/src/controller/home/controller.dart';
-
-export 'package:example/src/controller/home/another_controller.dart';
 
 export 'package:example/src/controller/home/yet_another_controller.dart';