+++ example/lib/src/view/home/home_page.dart
@@ -18,7 +18,6 @@ class HomePage extends StatefulWidget {
   State<StatefulWidget> createState() => _HomePageState();
 }
 
-/// Insert an InheritedWidget at this point to be used.
 class _HomePageState extends StateX<HomePage> {
   /// Let the 'business logic' run in a Controller
   _HomePageState() : super(Controller()) {
@@ -34,72 +33,82 @@ class _HomePageState extends StateX<HomePage> {
     /// all it's Controllers if any.
     super.initState();
 
+    var con = controller;
+
+    con = controllerByType<Controller>();
+
+    con = controllerById(con?.identifier);
+
     /// Retrieve the 'app level' State object
     appState = rootState!;
 
     /// You're able to retrieve the Controller(s) from other State objects.
-    var con = appState.controller;
+    con = appState.controller;
 
-    con = appState.controllerByType<AppController>();
+    con = appState.controllerByType<Controller>();
 
     con = appState.controllerById(con?.identifier);
+
+    /// You're able to retrieve the 'past' Controller(s)
+    con = controllerByType<AppController>();
+
+    con = appState.controllerByType<AppController>();
   }
 
   late AppStateX appState;
 
   /// Build the 'child' Widget passed to the InheritedWidget.
   @override
-  Widget buildWidget(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(widget.title),
-      ),
-      body: Center(
-        /// SetState class is like a setState() function but called only when the App's InheritedWidget is rebuilt.
-        child: SetState(
-          builder: (context, dataObject) => Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget>[
-              /// Display the App's data object if it has something to display
-              /// Note, the Stat Controllers has access to the same 'dataObject'
-              if (dataObject != null &&
-                  dataObject == con.dataObject &&
-                  con.dataObject is String)
-                Padding(
-                  padding: const EdgeInsets.all(30),
-                  child: Text(
-                    dataObject as String,
-                    key: const Key('greetings'),
-                    style: TextStyle(
-                      color: Colors.red,
-                      fontSize: Theme.of(context).textTheme.headline4!.fontSize,
+  Widget buildWidget(BuildContext context) => Scaffold(
+        appBar: AppBar(
+          title: Text(widget.title),
+        ),
+        body: Center(
+          /// SetState class is like a setState() function but called only when the App's InheritedWidget is rebuilt.
+          child: SetState(
+            builder: (context, dataObject) => Column(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: <Widget>[
+                /// Display the App's data object if it has something to display
+                /// Note, the Stat Controllers has access to the same 'dataObject'
+                if (dataObject != null &&
+                    dataObject == con.dataObject &&
+                    con.dataObject is String)
+                  Padding(
+                    padding: const EdgeInsets.all(30),
+                    child: Text(
+                      dataObject as String,
+                      key: const Key('greetings'),
+                      style: TextStyle(
+                        color: Colors.red,
+                        fontSize:
+                            Theme.of(context).textTheme.headline4!.fontSize,
+                      ),
                     ),
                   ),
+                Text(
+                  'You have pushed the button this many times:',
+                  style: Theme.of(context).textTheme.bodyText2,
                 ),
-              Text(
-                'You have pushed the button this many times:',
-                style: Theme.of(context).textTheme.bodyText2,
-              ),
-              Text(
-                '${con.count}',
-                style: Theme.of(context).textTheme.headline4,
-              ),
-            ],
+                Text(
+                  '${con.count}',
+                  style: Theme.of(context).textTheme.headline4,
+                ),
+              ],
+            ),
           ),
         ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        key: const Key('+'),
+        floatingActionButton: FloatingActionButton(
+          key: const Key('+'),
 
-        /// Refresh only the Text widget containing the counter.
-        onPressed: () => con.incrementCounter(),
+          /// Refresh only the Text widget containing the counter.
+          onPressed: () => con.incrementCounter(),
 
-        /// You can have the Controller called the interface (the View).
+          /// You can have the Controller called the interface (the View).
 //          onPressed: con.onPressed,
-        child: const Icon(Icons.add),
-      ),
-    );
-  }
+          child: const Icon(Icons.add),
+        ),
+      );
 
   /// Supply an error handler for Unit Testing.
   @override